refactor: step3 패키지 이름 변경, 테스트 수
