+++ src/main/java/racingcar/car/CarController.java
@@ -1,0 +1,24 @@
+package racingcar.car;
+
+import racingcar.random.RandomNumberGenerator;
+
+public class CarController {
+    public static final int ALLOWED_MINIMUM_DISTANCE = 4;
+
+    private final RandomNumberGenerator randomNumberGenerator;
+
+    public CarController(RandomNumberGenerator randomNumberGenerator) {
+        this.randomNumberGenerator = randomNumberGenerator;
+    }
+
+    public void moveCar(Car car) {
+        int randomDistance = randomNumberGenerator.generate();
+        if (canMove(randomDistance)) {
+            car.move();
+        }
+    }
+
+    private boolean canMove(int distance) {
+        return distance >= ALLOWED_MINIMUM_DISTANCE;
+    }
+}
