+++ src/main/java/racingcar/random/RandomNumberGenerator.java
@@ -1,0 +1,5 @@
+package racingcar.random;
+
+public interface RandomNumberGenerator {
+    int generate();
+}
