--- src/main/java/step2/Application.java
@@ -1,9 +1,0 @@
-package step2;
-
-import step2.delimiter.DelimiterNumberParserFactory;
-
-public class Application {
-    public static void main(String[] args) {
-        new StringAddCalculator(new DelimiterNumberParserFactory());
-    }
-}
