--- src/main/java/step2/delimiter/DelimiterNumberParserFactory.java
@@ -1,18 +1,0 @@
-package step2.delimiter;
-
-import step2.exception.StringAddIllegalArgumentException;
-
-public class DelimiterNumberParserFactory {
-    private final DefaultDelimiterNumberParser defaultDelimiter = new DefaultDelimiterNumberParser();
-    private final CustomDelimiterNumberParser customDelimiterNumberParser = new CustomDelimiterNumberParser();
-
-    public NumberParser getNumberParser(String text) {
-        if (customDelimiterNumberParser.hasDelimiter(text)) {
-            return customDelimiterNumberParser;
-        }
-        if (defaultDelimiter.hasDelimiter(text)) {
-            return defaultDelimiter;
-        }
-        throw StringAddIllegalArgumentException.DELIMITER_NOT_FOUND;
-    }
-}
