--- src/main/java/step2/delimiter/NumberParser.java
@@ -1,9 +1,0 @@
-package step2.delimiter;
-
-import java.util.List;
-
-public interface NumberParser {
-    List<String> split(String text);
-
-    boolean hasDelimiter(String text);
-}
