--- src/main/java/step3/random/RandomNumberGenerator.java
@@ -1,5 +1,0 @@
-package step3.random;
-
-public interface RandomNumberGenerator {
-    int generate();
-}
