+++ src/main/java/stringadder/Application.java
@@ -1,0 +1,9 @@
+package stringadder;
+
+import stringadder.delimiter.DelimiterNumberParserFactory;
+
+public class Application {
+    public static void main(String[] args) {
+        new StringAddCalculator(new DelimiterNumberParserFactory());
+    }
+}
