+++ src/main/java/stringadder/delimiter/NumberParser.java
@@ -1,0 +1,9 @@
+package stringadder.delimiter;
+
+import java.util.List;
+
+public interface NumberParser {
+    List<String> split(String text);
+
+    boolean hasDelimiter(String text);
+}
